nitypes: Move type aliases and type variables to a public submodule so it's less prominent in autocomplete
