+++ src/nitypes/waveform/_digital/__init__.py
@@ -3,11 +3,6 @@
 from nitypes.waveform._digital._signal import DigitalWaveformSignal
 from nitypes.waveform._digital._signal_collection import DigitalWaveformSignalCollection
 from nitypes.waveform._digital._state import DigitalState
-from nitypes.waveform._digital._types import (
-    AnyDigitalState,
-    TDigitalState,
-    TOtherDigitalState,
-)
 from nitypes.waveform._digital._waveform import (
     DigitalWaveform,
     DigitalWaveformFailure,
@@ -15,13 +10,10 @@ from nitypes.waveform._digital._waveform import (
 )
 
 __all__ = [
-    "AnyDigitalState",
     "DigitalState",
     "DigitalWaveform",
     "DigitalWaveformFailure",
     "DigitalWaveformSignal",
     "DigitalWaveformSignalCollection",
     "DigitalWaveformTestResult",
-    "TDigitalState",
-    "TOtherDigitalState",
 ]
