+++ src/nitypes/waveform/_timing/__init__.py
@@ -2,28 +2,8 @@
 
 from nitypes.waveform._timing._sample_interval import SampleIntervalMode
 from nitypes.waveform._timing._timing import Timing
-from nitypes.waveform._timing._types import (
-    TOtherSampleInterval,
-    TOtherTimeOffset,
-    TOtherTimestamp,
-    TSampleInterval,
-    TSampleInterval_co,
-    TTimeOffset,
-    TTimeOffset_co,
-    TTimestamp,
-    TTimestamp_co,
-)
 
 __all__ = [
     "SampleIntervalMode",
     "Timing",
-    "TOtherSampleInterval",
-    "TOtherTimeOffset",
-    "TOtherTimestamp",
-    "TSampleInterval_co",
-    "TSampleInterval",
-    "TTimeOffset_co",
-    "TTimeOffset",
-    "TTimestamp_co",
-    "TTimestamp",
 ]
